requirements.txt
selenium==4.22.0
undetected-chromedriver==3.5.5
chromedriver_autoinstaller
python-dotenv
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
pydantic==2.8.2
requests==2.31.0
openai
numpy
tweepy
eth_keys
web3
twitter-api-client
twitter_openapi_python
